refactor(providers): enforce provider boundaries
