+++ cmd/usagebar/main.go
@@ -12,9 +12,9 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/senna-lang/herdr-agent-usage/internal/claude"
 	"github.com/senna-lang/herdr-agent-usage/internal/herdrcli"
 	"github.com/senna-lang/herdr-agent-usage/internal/limits"
+	"github.com/senna-lang/herdr-agent-usage/internal/providers/claude"
 	"github.com/senna-lang/herdr-agent-usage/internal/ratelimit"
 	"github.com/senna-lang/herdr-agent-usage/internal/setup"
 	"github.com/senna-lang/herdr-agent-usage/internal/update"
