+++ internal/limits/primarynotify.go
@@ -7,7 +7,7 @@
 package limits
 
 import (
-	"github.com/senna-lang/herdr-agent-usage/internal/claude"
+	"github.com/senna-lang/herdr-agent-usage/internal/providers/claude"
 	"github.com/senna-lang/herdr-agent-usage/internal/ratelimit"
 )
