+++ internal/limits/types.go
@@ -1,62 +1,20 @@
 /**
  * Per-provider rate-limit snapshot used for display.
+ *
+ * The concrete definitions live in internal/limitscore, one layer below both
+ * internal/limits and every provider package, since provider adapters must
+ * construct ProviderLimits without importing internal/limits (see
+ * internal/limitscore/types.go). These aliases keep every existing call site
+ * in this package unqualified.
  */
 package limits
 
-// RunOutEstimate projects when a window will be exhausted at the recently
-// observed consumption pace (approach B). Attached to a window after the
-// history pass; the formatter only reads it.
-type RunOutEstimate struct {
-	// MinutesToEmpty until empty. 0 means already empty.
-	MinutesToEmpty float64 `json:"minutesToEmpty"`
-	// EmptyBeforeReset is true when projected to empty before its reset (only then warn).
-	EmptyBeforeReset bool `json:"emptyBeforeReset"`
-}
+import "github.com/senna-lang/herdr-agent-usage/internal/limitscore"
 
-// LimitWindow is one rate-limit window snapshot.
-type LimitWindow struct {
-	// UsedPercentage (0-100). Remaining display = 100 - used.
-	UsedPercentage float64 `json:"usedPercentage"`
-	// ResetsAt is unix epoch seconds. Nil when unknown.
-	ResetsAt *int64 `json:"resetsAt,omitempty"`
-	// WindowMinutes hints the window length for display labels.
-	WindowMinutes *int `json:"windowMinutes,omitempty"`
-	// RunOut projection from the recent-pace history pass. Nil when there is
-	// not enough history, the pace is flat/negative, or it holds until reset.
-	RunOut *RunOutEstimate `json:"runOut,omitempty"`
-}
+type RunOutEstimate = limitscore.RunOutEstimate
 
-// ProviderLimits is a per-provider rate-limit snapshot used for display.
-type ProviderLimits struct {
-	ProviderID string
-	Label      string
-	// Primary is the short window (5h / session).
-	Primary *LimitWindow
-	// Secondary is the mid window (weekly).
-	Secondary *LimitWindow
-	// Tertiary is the long window (monthly). OpenCode Go etc.
-	Tertiary    *LimitWindow
-	PlanType    *string
-	Source      string
-	FetchedAtMs int64
-	Note        *string
-	// PaneActivity is per-pane token activity share over the smallest window.
-	PaneActivity *ProviderPaneActivity
-	// GroupLabel nests this entry under one shared heading in the panel
-	// (e.g. multiple configured Claude accounts under "Claude") instead of
-	// its own top-level block. Only takes effect when 2+ entries share the
-	// same non-empty GroupLabel; a lone entry renders as if it were empty.
-	GroupLabel string
-	// AccountLabel is shown in place of Label for the per-account line inside
-	// a group (e.g. the account's real login email), so members sharing one
-	// GroupLabel stay distinguishable. Ignored outside a group.
-	AccountLabel string
-}
+type LimitWindow = limitscore.LimitWindow
 
-// ProviderPaneActivity is windowed per-pane activity for one provider.
-type ProviderPaneActivity struct {
-	// WindowMinutes used for the aggregation.
-	WindowMinutes int
-	TotalTokens   int
-	Panes         []PaneActivityShare
-}
+type ProviderLimits = limitscore.ProviderLimits
+
+type ProviderPaneActivity = limitscore.ProviderPaneActivity
