+++ internal/providers/claude/modelwindows.go
