+++ internal/providers/claude/modelwindows_test.go
