+++ internal/providers/claude/profile.go
