+++ internal/providers/claude/profile_test.go
